python_scripts/requirements.txt
functions-framework
solana
google-cloud-secret-manager